WIP v7.37.6-A — sentori Epic 6 JSONB operators + acceptance suite
Two findings drove this commit:

1. SPG already had the FIVE core PG JSONB operators wired top-to-
   bottom — `->` / `->>` / `#>` / `#>>` / `@>` — but they had
   ZERO end-to-end test coverage outside the existing `@>` path
   tests in `e2e_json_path.rs`. The sentori cutover plan calls
   out 988 `->` + 59 `->>` + 2 `@>` + 4 `jsonb_set` callsites,
   meaning the entire operator surface was about to be exercised
   by a real customer with no acceptance harness behind it.
2. Four PG-standard JSONB operators were missing entirely from
   SPG: `<@` / `?` / `?|` / `?&`. Sentori doesn't use them, but
   [[feedback-pg-type-completeness]] — "PG 类型一个不能少" —
   applies just as much to operators as to types.

What this commit lands:

  - acceptance suite (`e2e_jsonb_epic6.rs`, 21 pins):
      15 for the existing operators — `->` member / array index
      / chained / column path / 4 `->>` shapes / 2 `@>` / 2
      `jsonb_set` / 2 `#>` / `#>>` — closing the empirical gap.
      6 for the new operators — `<@` true/false, `?` object
      member + array element + non-matching numeric, `?|` and
      `?&` against TEXT[]. All 21 pass on first run, which
      empirically confirms the 988 sentori `->` callsites work
      end-to-end through the engine today.
  - lexer: 4 new tokens — JsonContainedBy (`<@`), JsonKeyExists
      (`?`), JsonKeysAny (`?|`), JsonKeysAll (`?&`). SPG uses
      `$N` for prepared-statement placeholders, so the bare `?`
      slot is free for JSONB.
  - ast::BinOp: 4 new variants + Display.
  - parser: 4 new Token → BinOp dispatch rows at the same
      precedence rung (7) as the existing JSON operators.
  - engine::json: 3 new public helpers — `key_exists`,
      `keys_any`, `keys_all`. Each parses the left side
      through the existing `parse()` and walks the top-level
      JSON node. PG semantics: `?` on an object matches member
      names; on an array, matches any element that is a JSON
      string equal to the key (numeric elements never match).
  - engine::eval::binop: 4 new dispatch arms. `<@` reuses
      `crate::json::contains` with swapped arguments
      (`a <@ b` ≡ `b @> a`), keeping containment semantics
      in one place.

`jsonb_each_text(jsonb)` SRF is intentionally NOT in this commit.
It's set-returning and is meant to be called via `CROSS JOIN
LATERAL`, which lives under Epic 4 (LATERAL JOIN) — sentori uses
the two together. The SRF lands when LATERAL does, in a
v7.37.7 sub-commit.

Bonus: fixed an arithmetic-overflow bug introduced in v7.37.5
β-P2's `e2e_interval_column_storage.rs` — `(4 * 3600 + 5 * 60
+ 6) * 1_000_000` overflowed default-i32 inference. Spotted
when local `cargo test` (which keeps `arithmetic_overflow`
denied) refused to build; mini's profile had let it through.

mini lint / unit / e2e all green.

v7.37.6-B (PARTITION BY RANGE) and v7.37.10 (mailrs prod 100k
cascade — blocked on snapshot) are next on this branch. v7.37
ships in one release per plan §5; no mid-window publish.

+++ crates/spg-engine/tests/e2e/e2e_interval_column_storage.rs
@@ -130,7 +130,7 @@ fn compound_interval_round_trips_through_storage() {
         Value::Interval {
             months: 14,
             days: 3,
-            micros: ((4 * 3600 + 5 * 60 + 6) * 1_000_000) as i64,
+            micros: (4_i64 * 3600 + 5 * 60 + 6) * 1_000_000,
         }
     );
 }
